requirements.txt
# NLP_phishing_detection_Project
# Auteur: Souleymane Sall
# Email: sallsouleymane2207@gmail.com

# ============================================
# Core ML/NLP
# ============================================
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1

# ============================================
# NLP & Feature Engineering
# ============================================
nltk==3.9.1

# ============================================
# Hyperparameter Optimization
# ============================================
optuna==3.6.1

# ============================================
# Experiment Tracking
# ============================================
mlflow==2.16.2

# ============================================
# Configuration Management
# ============================================
hydra-core==1.3.2
omegaconf==2.3.0

# ============================================
# API & Serving
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
python-multipart==0.0.12

# ============================================
# User Interface
# ============================================
streamlit==1.39.0

# ============================================
# Data Acquisition
# ============================================
kaggle==1.6.17

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.1
joblib==1.4.2
tqdm==4.66.5

# ============================================
# Visualization
# ============================================
matplotlib==3.9.2
seaborn==0.13.2

# ============================================
# Testing
# ============================================
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2

# ============================================
# Code Quality
# ============================================
ruff==0.6.9
black==24.10.0
isort==5.13.2
mypy==1.11.2

# ============================================
# Type Stubs
# ============================================
types-requests==2.32.0.20240914
pandas-stubs==2.2.2.240909